+++ InstantMock/Arguments/Arg.swift
@@ -11,6 +11,8 @@
 public class Arg<T> {
 
 
+    // MARK: Arguments matching exact values
+
     /** Register a mandatory value */
     public static func eq(_ val: T) -> T {
         let factory = ArgumentFactoryImpl<T>()
@@ -45,28 +47,51 @@ public class Arg<T> {
     }
 
 
-    /** Register a closure to be verified */
+    // MARK: Arguments matching a condition
+
+    /** Register a closure to be verified, with mandatory type */
     public static func verify(_ condition: @escaping (T) -> Bool) -> T {
         let factory = ArgumentFactoryImpl<T>()
         return Arg.verify(condition, argFactory: factory, argStorage: ArgumentStorageImpl.instance)
     }
 
 
-    /** Register a closure to be verified */
+    /** Register a closure to be verified, with optional type */
+    public static func verify(_ condition: @escaping (T?) -> Bool) -> T? {
+        let factory = ArgumentFactoryImpl<T>()
+        return Arg.verify(condition, argFactory: factory, argStorage: ArgumentStorageImpl.instance)
+    }
+
+
+    /** Register a closure to be verified, with mandatory type */
     static func verify<F>(_ condition: @escaping (T) -> Bool, argFactory: F, argStorage: ArgumentStorage) -> T
         where F: ArgumentFactory, F.Value == T {
 
-        // create and store instance
-        let arg = argFactory.argument(condition: condition)
-        argStorage.store(arg)
+            let arg = argFactory.argument(condition: condition)
+            argStorage.store(arg)
 
-        // return default value
-        guard let ret = DefaultValueHandler<T>().it else {
-            fatalError("Unexpected type, only `MockUsable` types can be used with `verify`")
-        }
-        return ret
+            // return default value
+            guard let ret = DefaultValueHandler<T>().it else {
+                fatalError("Unexpected type, only `MockUsable` types can be used with `verify`")
+            }
+            return ret
     }
 
+
+    /** Register a closure to be verified, with optional type */
+    static func verify<F>(_ condition: @escaping (T?) -> Bool, argFactory: F, argStorage: ArgumentStorage) -> T?
+        where F: ArgumentFactory, F.Value == T {
+
+            // create and store instance
+            let arg = argFactory.argument(condition: condition)
+            argStorage.store(arg)
+
+            // return default nil value
+            return nil
+    }
+
+
+    // MARK: Arguments matching any values
 
     /** Register any value */
     public static var any: T {