+++ InstantMock/Arguments/Helpers/ArgumentFactory.swift
@@ -19,6 +19,7 @@ public protocol ArgumentFactory {
 
     // Create a new argument that must verify provided condition
     func argument(condition: @escaping (Value) -> Bool) -> ArgumentVerify
+    func argument(condition: @escaping (Value?) -> Bool) -> ArgumentVerify
 
     // Create a new argument capture
     func argumentCapture(_ typeDescription: String) -> ArgumentCapture
@@ -38,7 +39,11 @@ class ArgumentFactoryImpl<T>: ArgumentFactory {
     }
 
     func argument(condition: @escaping (T) -> Bool) -> ArgumentVerify {
-        return ArgumentVerifyImpl<T>(condition)
+        return ArgumentVerifyMandatoryImpl<T>(condition)
+    }
+
+    func argument(condition: @escaping (T?) -> Bool) -> ArgumentVerify {
+        return ArgumentVerifyOptionalImpl<T>(condition)
     }
 
     func argumentCapture(_ typeDescription: String) -> ArgumentCapture {