+++ InstantMockTests/Unit Tests/Arguments/Helpers/ArgumentsConfigurationTests.swift
@@ -40,10 +40,10 @@ class ArgumentsConfigurationTests: XCTestCase {
 
 
     func testEquality() {
-        let args1 = ArgumentsConfiguration([ArgumentValueMock(12), ArgumentVerifyMock<String>({str in true})])
+        let args1 = ArgumentsConfiguration([ArgumentValueMock(12), ArgumentVerifyMandatoryMock<String>({str in true})])
         XCTAssertEqual(args1, args1)
 
-        var args2 = ArgumentsConfiguration([ArgumentValueMock(13), ArgumentVerifyMock<String>({str in true})])
+        var args2 = ArgumentsConfiguration([ArgumentValueMock(13), ArgumentVerifyMandatoryMock<String>({str in true})])
         XCTAssertEqual(args1, args2)
 
         args2 = ArgumentsConfiguration([ArgumentValueMock(13), ArgumentValueMock(14)])
@@ -52,26 +52,26 @@ class ArgumentsConfigurationTests: XCTestCase {
         args2 = ArgumentsConfiguration([ArgumentValueMock(13), ArgumentAnyMock()])
         XCTAssertNotEqual(args1, args2)
 
-        args2 = ArgumentsConfiguration([ArgumentVerifyMock<String>({str in false}), ArgumentVerifyMock<String>({str in true})])
+        args2 = ArgumentsConfiguration([ArgumentVerifyMandatoryMock<String>({str in false}), ArgumentVerifyMandatoryMock<String>({str in true})])
         XCTAssertNotEqual(args1, args2)
 
-        args2 = ArgumentsConfiguration([ArgumentVerifyMock<String>({str in false}), ArgumentAnyMock()])
+        args2 = ArgumentsConfiguration([ArgumentVerifyMandatoryMock<String>({str in false}), ArgumentAnyMock()])
         XCTAssertNotEqual(args1, args2)
     }
 
 
     func testGreaterThan() {
-        let args1 = ArgumentsConfiguration([ArgumentValueMock(12), ArgumentVerifyMock<String>({str in true})])
+        let args1 = ArgumentsConfiguration([ArgumentValueMock(12), ArgumentVerifyMandatoryMock<String>({str in true})])
         var args2 = ArgumentsConfiguration([ArgumentValueMock(13), ArgumentValueMock(14)])
         XCTAssertGreaterThan(args2, args1)
 
         args2 = ArgumentsConfiguration([ArgumentValueMock(13), ArgumentAnyMock()])
         XCTAssertGreaterThan(args1, args2)
 
-        args2 = ArgumentsConfiguration([ArgumentVerifyMock<String>({str in false}), ArgumentVerifyMock<String>({str in true})])
+        args2 = ArgumentsConfiguration([ArgumentVerifyMandatoryMock<String>({str in false}), ArgumentVerifyMandatoryMock<String>({str in true})])
         XCTAssertGreaterThan(args1, args2)
 
-        args2 = ArgumentsConfiguration([ArgumentVerifyMock<String>({str in false}), ArgumentAnyMock()])
+        args2 = ArgumentsConfiguration([ArgumentVerifyMandatoryMock<String>({str in false}), ArgumentAnyMock()])
         XCTAssertGreaterThan(args1, args2)
     }